Reader: A/B various height limits on featured images.
Adds in a new A/B test that can try out various limits on featured image height. Doesn't affect embeds, as they're much harder to work with and predict with max heights like these.

Also changes A/B test a bit to allow folks that are already in an A/B test to continue being in that test, even if they're not eligible. This lets us debug A/B tests that most devs are not eligible for.