update to treat local response as provisional data
This is a biggish commit, but the crux is to change how we process local and server responses so that new post-lists from a local response are treated as temporary and are re-evaluted when we get the server response for that request.

This has the corner case potential to get out of whack if/when the server responses come in out of order. If that happens, we just wipe the cached list clean so the app has the ability to rebuild the list from scratch.

I am also renaming some variables to be a little more explicit, like `listKey` instead of `key` and `requestKey`, and `newPosts` and `removedPosts`, etc.